pyproject.toml
[virtualenvs]
in-project = true

[tool.poetry]
name = "fog-computing-prototype"
version = "0.1.0"
description = ""
authors = ["filiprejmus <rejmusfilip@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.9"
pyzmq = "25.0.1"
numpy = "^1.25.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"